+++ db/migrate/003_create_scores.rb
@@ -17,6 +17,7 @@ class CreateScores < ActiveRecord::Migration[5.1]
       t.float :tes
       t.float :pcs
       t.float :deductions
+      t.string :deduction_reasons
       t.float :base_value
       t.string :elements_summary
       t.string :components_summary